bring new file structure to the dev branch (#11)
* Clean up index.js / create app structure

broke up the clutter that was the original index.js by created three main files: tools, bot, and input.

In order to do this, also created a micro back end with node express to host the static html and expose the public folder to the client.

* updated main.css

looks like this was an older version of the style file, so ill bring it up to date before we merge this with dev.

* Update README.md

* Revert "updated index.html"

This reverts commit 9efc9ddcf50f91b5bab4d643c2c1a6087b53daa5, reversing
changes made to a335749d232977559e27bcfa2b8651a5e24fe12f.

* updated index.html
