+++ README.md
@@ -2,6 +2,13 @@
 
 ApostleAI is a conversation bot modeled after Paul the Apostle, commonly known as Saint Paul, an apostle who spread the teachings of Jesus in the first-century world.
 
-[View the live website here!](https://prophet-ai.github.io/ApostleAI/)
+To run the this project on your local machine;
+1. Install [`NodeJS`](https://nodejs.org/en/download/)
+2. In your terminal, navigate to the project directory and hit it with a `npm install` which should install all dependencies, in this case the only one necessary should be [Express.js](https://github.com/expressjs/express).
+
+Once the dependencies have been installed, to start the local server;
+1. In your terminal type `npm run start` which will run the server and host the application.
+2. Nagivate to `localhost:8080` to see the project live. 
+> Note: any updates to the files in your project will automatically be dispalyed on this server.
 
 ![alt text](https://upload.wikimedia.org/wikipedia/commons/thumb/6/67/The_Predication_of_Saint_Paul_LACMA_M.2000.179.24.jpg/800px-The_Predication_of_Saint_Paul_LACMA_M.2000.179.24.jpg)
